Macro feedback + remove resource id on PartitionProcessorRpcRequest
