+++ crates/types/src/id_util.rs
@@ -46,7 +46,6 @@ prefixed_ids! {
         Subscription("sub"),
         Awakeable("prom"),
         Snapshot("snap"),
-        PartitionProcessorRpcRequest("rpcid"),
     }
 }
 
@@ -229,124 +228,4 @@ impl<T: ResourceId + ?Sized> IdEncoder<T> {
     pub fn finalize(self) -> String {
         self.buf
     }
-}
-
-#[macro_export]
-macro_rules! uuid_backed_id {
-    ($name:ident) => {
-        paste::paste! {
-            uuid_backed_id!(@gen [< $name Id >], $name);
-        }
-    };
-    (@gen $id_name:ident, $resource_type:ident) => {
-        #[derive(
-            PartialEq,
-            Eq,
-            Clone,
-            Copy,
-            Hash,
-            PartialOrd,
-            Ord,
-            serde_with::SerializeDisplay,
-            serde_with::DeserializeFromStr,
-        )]
-        pub struct $id_name(pub(crate) Ulid);
-
-        impl $id_name {
-            pub fn new() -> Self {
-                Self(Ulid::new())
-            }
-
-            pub const fn from_parts(timestamp_ms: u64, random: u128) -> Self {
-                Self(Ulid::from_parts(timestamp_ms, random))
-            }
-
-            pub fn from_slice(b: &[u8]) -> Result<Self, IdDecodeError> {
-                let ulid = Ulid::from_bytes(b.try_into().map_err(|_| IdDecodeError::Length)?);
-                debug_assert!(!ulid.is_nil());
-                Ok(Self(ulid))
-            }
-
-            pub fn from_bytes(bytes: [u8; 16]) -> Self {
-                let ulid = Ulid::from_bytes(bytes);
-                debug_assert!(!ulid.is_nil());
-                Self(ulid)
-            }
-
-            pub fn to_bytes(&self) -> [u8; 16] {
-                self.0.to_bytes()
-            }
-        }
-
-        impl Default for $id_name {
-            fn default() -> Self {
-                Self::new()
-            }
-        }
-
-        impl ResourceId for $id_name {
-            const SIZE_IN_BYTES: usize = size_of::<Ulid>();
-            const RESOURCE_TYPE: IdResourceType = IdResourceType::$resource_type;
-            const STRING_CAPACITY_HINT: usize = base62_max_length_for_type::<u128>();
-
-            fn push_contents_to_encoder(&self, encoder: &mut IdEncoder<Self>) {
-                let raw: u128 = self.0.into();
-                encoder.encode_fixed_width(raw);
-            }
-        }
-
-        impl TimestampAwareId for $id_name {
-            fn timestamp(&self) -> MillisSinceEpoch {
-                self.0.timestamp_ms().into()
-            }
-        }
-
-        impl FromStr for $id_name {
-            type Err = IdDecodeError;
-
-            fn from_str(input: &str) -> Result<Self, Self::Err> {
-                let mut decoder = IdDecoder::new(input)?;
-                // Ensure we are decoding the correct resource type
-                if decoder.resource_type != Self::RESOURCE_TYPE {
-                    return Err(IdDecodeError::TypeMismatch);
-                }
-
-                // ulid (u128)
-                let raw_ulid: u128 = decoder.cursor.decode_next()?;
-                Ok(Self::from(raw_ulid))
-            }
-        }
-
-        impl fmt::Display for $id_name {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                let mut encoder = IdEncoder::<Self>::new();
-                self.push_contents_to_encoder(&mut encoder);
-                fmt::Display::fmt(&encoder.finalize(), f)
-            }
-        }
-
-        impl fmt::Debug for $id_name {
-            fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-                // use the same formatting for debug and display to show a consistent representation
-                fmt::Display::fmt(self, f)
-            }
-        }
-
-        impl From<u128> for $id_name {
-            fn from(value: u128) -> Self {
-                Self(Ulid::from(value))
-            }
-        }
-
-        #[cfg(feature = "schemars")]
-        impl schemars::JsonSchema for $id_name {
-            fn schema_name() -> String {
-                <String as schemars::JsonSchema>::schema_name()
-            }
-
-            fn json_schema(g: &mut schemars::gen::SchemaGenerator) -> schemars::schema::Schema {
-                <String as schemars::JsonSchema>::json_schema(g)
-            }
-        }
-    };
 }
